requirements.txt
absl-py==1.4.0
adjustText==1.1.1
anndata==0.8.0
anyio==3.7.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
cfgv==3.4.0
charset-normalizer==3.2.0
cmcrameri==1.9
colorama==0.4.6
comm==0.1.4
contourpy==1.1.0
cramjam==2.7.0
csbdeep==0.7.4
cycler==0.11.0
debugpy==1.6.7.post1
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.7
distlib==0.3.7
EMD-signal==1.5.1
EntropyHub==0.2
exceptiongroup==1.1.2
executing==1.2.0
factor-analyzer==0.5.0
fastjsonschema==2.18.0
fastparquet==2023.7.0
fastremap==1.13.5
fastrlock==0.8.1
filelock==3.12.2
flatbuffers==23.5.26
fonttools==4.42.0
fqdn==1.5.1
fsspec==2023.6.0
gast==0.4.0
google-api-core==2.11.1
google-auth==2.21.0
google-auth-oauthlib==0.4.6
google-cloud-core==2.3.3
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
GPUtil==1.4.0
grpcio==1.57.0
h5py==3.11.0
identify==2.5.26
idna==3.4
imageio==2.31.3
importlib-metadata==6.8.0
importlib-resources==6.0.1
ipykernel==6.25.1
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.1.0
isoduration==20.11.0
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
jupyterlab_server==2.24.0
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
kneed==0.8.5
lazy_loader==0.3
libclang==16.0.0
llvmlite==0.40.1
looseversion==1.1.2
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.1
ml-dtypes==0.2.0
mudata==0.2.2
multiprocess==0.70.15
muon==0.1.3
namex==0.0.8
natsort==8.3.1
nbclient==0.8.0
nbconvert==7.7.3
nbformat==5.9.2
nest-asyncio==1.5.7
networkx==3.1
nodeenv==1.8.0
notebook==7.0.2
notebook_shim==0.2.3
numba==0.57.1
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.7.0.68
opencv-python-headless==4.8.0.74
opt-einsum==3.3.0
optree==0.11.0
overrides==7.4.0
packaging==23.1
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
pathos==0.3.1
patsy==0.5.3
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.10.0
plotly==5.16.0
pox==0.3.3
ppft==1.7.6.7
pre-commit==3.3.3
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==3.19.6
psutil==5.9.5
pure-eval==0.2.2
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydot==1.4.2
Pygments==2.16.1
pynndescent==0.5.8
pyparsing==3.0.9
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
pyqtgraph==0.13.3
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3
PyWavelets==1.4.1
pywin32==306
pywinpty==2.0.11
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.4.3
QtPy==2.3.1
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
roifile==2023.5.12
rpds-py==0.9.2
rsa==4.9
scanpy==1.9.3
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.11.1
seaborn==0.12.2
Send2Trash==1.8.2
session-info==1.0.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
stack-data==0.6.2
stardist==0.8.5
statannot==0.2.3
statsmodels==0.14.0
stdlib-list==0.8.0
tenacity==8.2.3
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.10.0
tensorflow-gpu==2.10.0
tensorflow-intel==2.14.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
terminado==0.17.1
threadpoolctl==3.2.0
tifffile==2023.8.30
tinycss2==1.2.1
tomli==2.0.1
torchaudio==2.0.2+cu118
torchvision==0.15.2+cu118
tornado==6.3.3
tqdm==4.64.1
trackpy==0.6.1
traitlets==5.9.0
typing_extensions==4.7.1
tzdata==2023.3
umap-learn==0.5.3
uri-template==1.3.0
urllib3==1.26.16
virtualenv==20.24.3
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
Werkzeug==2.3.6
widgetsnbextension==4.0.8
wrapt==1.14.1
zipp==3.16.2
